+++ minifica.data/IRepository/IActionRepository.cs
@@ -1,0 +1,7 @@
+﻿namespace minifica.data.IRepository
+{
+    public interface IActionRepository : IBaseRepository<Models.Action>
+    {
+
+    }
+}