+++ minifica.data/Models/Country.cs
@@ -1,12 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace minifica.data.Models
+﻿namespace minifica.data.Models
 {
-	public class Country : BaseEntity
+    public class Country : BaseEntity
 	{
 		public string Name { get; set; }
 	}