+++ minifica.api/CHANGELOG.md
@@ -3,6 +3,12 @@
 ## [DevOps-4] - [2023-06-02] - Miguel Lopez
 * Created CRUD entity Country.
 
+## [DevOps-10] - [2023-06-05] - Alfredo Ferretis
+* Created CRUD entity Profile.
+
+## [DevOps-9] - [2023-06-05] - Alfredo Ferretis
+* Created CRUD entity Action.
+
 ## [DevOps-8] - [2023-06-02] - Alfredo Ferretis
 * Created CRUD entity Module.